#2 Commit

+++ src/main/java/org/udv/nrc/caixaudv/web/rest/CompraResource.java
@@ -1,20 +1,35 @@
 package org.udv.nrc.caixaudv.web.rest;
-import org.udv.nrc.caixaudv.domain.Compra;
-import org.udv.nrc.caixaudv.repository.CompraRepository;
-import org.udv.nrc.caixaudv.web.rest.errors.BadRequestAlertException;
-import org.udv.nrc.caixaudv.web.rest.util.HeaderUtil;
-import io.github.jhipster.web.util.ResponseUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
-
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+
+import javax.validation.Valid;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.udv.nrc.caixaudv.domain.Compra;
+import org.udv.nrc.caixaudv.domain.Conta;
+import org.udv.nrc.caixaudv.domain.enumeration.NivelPermissao;
+import org.udv.nrc.caixaudv.repository.CompraRepository;
+import org.udv.nrc.caixaudv.repository.ContaRepository;
+import org.udv.nrc.caixaudv.security.UserAccountPermissionChecker;
+import org.udv.nrc.caixaudv.security.UserNotActivatedException;
+import org.udv.nrc.caixaudv.web.rest.errors.BadRequestAlertException;
+import org.udv.nrc.caixaudv.web.rest.errors.UserNotAuthorizedException;
+import org.udv.nrc.caixaudv.web.rest.util.HeaderUtil;
 
 /**
  * REST controller for managing Compra.
@@ -28,6 +43,12 @@ public class CompraResource {
     private static final String ENTITY_NAME = "compra";
 
     private final CompraRepository compraRepository;
+
+    @Autowired
+    private ContaRepository contaRepository;
+
+    private final List<NivelPermissao> canCRDAll = Arrays.asList(NivelPermissao.ADMIN, 
+        NivelPermissao.OPERADOR);
 
     public CompraResource(CompraRepository compraRepository) {
         this.compraRepository = compraRepository;
@@ -46,10 +67,14 @@ public class CompraResource {
         if (compra.getId() != null) {
             throw new BadRequestAlertException("A new compra cannot already have an ID", ENTITY_NAME, "idexists");
         }
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!UserAccountPermissionChecker.checkPermissao(contaTest, canCRDAll)){
+            throw new UserNotAuthorizedException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         Compra result = compraRepository.save(compra);
-        return ResponseEntity.created(new URI("/api/compras/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
-            .body(result);
+            return ResponseEntity.created(new URI("/api/compras/" + result.getId()))
+                .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
+                .body(result);
     }
 
     /**
@@ -67,6 +92,10 @@ public class CompraResource {
         if (compra.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!contaTest.getNivelPermissao().equals(NivelPermissao.ADMIN)) {
+            throw new UserNotAuthorizedException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         Compra result = compraRepository.save(compra);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, compra.getId().toString()))
@@ -80,8 +109,15 @@ public class CompraResource {
      */
     @GetMapping("/compras")
     public List<Compra> getAllCompras() {
-        log.debug("REST request to get all Compras");
-        return compraRepository.findAll();
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        boolean isClient = contaTest.getNivelPermissao().equals(NivelPermissao.CLIENTE);
+        if(isClient){
+            return compraRepository.findByUserIsCurrentUser();
+        } 
+        else if(!isClient && !UserAccountPermissionChecker.checkPermissao(contaTest, canCRDAll)){
+            throw new UserNotAuthorizedException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
+        else return compraRepository.findAll();
     }
 
     /**
@@ -93,8 +129,19 @@ public class CompraResource {
     @GetMapping("/compras/{id}")
     public ResponseEntity<Compra> getCompra(@PathVariable Long id) {
         log.debug("REST request to get Compra : {}", id);
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!UserAccountPermissionChecker.checkPermissao(contaTest, canCRDAll)){
+            throw new UserNotAuthorizedException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         Optional<Compra> compra = compraRepository.findById(id);
-        return ResponseUtil.wrapOrNotFound(compra);
+        if(compra.isPresent()){
+            if(compra.get().getConta().equals(contaTest) || 
+                    contaTest.getNivelPermissao().equals(NivelPermissao.ADMIN) ||
+                    contaTest.getNivelPermissao().equals(NivelPermissao.OPERADOR))
+                return ResponseEntity.ok(compra.get());
+            else throw new UserNotAuthorizedException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
+        else return ResponseEntity.notFound().build();
     }
 
     /**
@@ -106,6 +153,10 @@ public class CompraResource {
     @DeleteMapping("/compras/{id}")
     public ResponseEntity<Void> deleteCompra(@PathVariable Long id) {
         log.debug("REST request to delete Compra : {}", id);
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!UserAccountPermissionChecker.checkPermissao(contaTest, canCRDAll)){
+            throw new UserNotAuthorizedException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         compraRepository.deleteById(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
