+++ src/main/java/org/udv/nrc/caixaudv/web/rest/ProdutoResource.java
@@ -1,18 +1,25 @@
 package org.udv.nrc.caixaudv.web.rest;
+import org.udv.nrc.caixaudv.domain.Conta;
 import org.udv.nrc.caixaudv.domain.Produto;
+import org.udv.nrc.caixaudv.domain.enumeration.NivelPermissao;
+import org.udv.nrc.caixaudv.repository.ContaRepository;
 import org.udv.nrc.caixaudv.repository.ProdutoRepository;
+import org.udv.nrc.caixaudv.security.UserAccountPermissionChecker;
 import org.udv.nrc.caixaudv.web.rest.errors.BadRequestAlertException;
 import org.udv.nrc.caixaudv.web.rest.util.HeaderUtil;
+
 import io.github.jhipster.web.util.ResponseUtil;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
-
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,6 +35,12 @@ public class ProdutoResource {
     private static final String ENTITY_NAME = "produto";
 
     private final ProdutoRepository produtoRepository;
+
+    @Autowired
+    private ContaRepository contaRepository;
+
+    private final List<NivelPermissao> canCRUDAll = Arrays.asList(NivelPermissao.ADMIN,
+        NivelPermissao.OPERADOR);
 
     public ProdutoResource(ProdutoRepository produtoRepository) {
         this.produtoRepository = produtoRepository;
@@ -46,10 +59,14 @@ public class ProdutoResource {
         if (produto.getId() != null) {
             throw new BadRequestAlertException("A new produto cannot already have an ID", ENTITY_NAME, "idexists");
         }
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!contaTest.getNivelPermissao().equals(NivelPermissao.ADMIN)){
+            throw new BadRequestAlertException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         Produto result = produtoRepository.save(produto);
-        return ResponseEntity.created(new URI("/api/produtos/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
-            .body(result);
+            return ResponseEntity.created(new URI("/api/produtos/" + result.getId()))
+                .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
+                .body(result);
     }
 
     /**
@@ -67,6 +84,10 @@ public class ProdutoResource {
         if (produto.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!contaTest.getNivelPermissao().equals(NivelPermissao.ADMIN)){
+            throw new BadRequestAlertException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         Produto result = produtoRepository.save(produto);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, produto.getId().toString()))
@@ -81,6 +102,10 @@ public class ProdutoResource {
     @GetMapping("/produtos")
     public List<Produto> getAllProdutos() {
         log.debug("REST request to get all Produtos");
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(UserAccountPermissionChecker.checkPermissao(contaTest, canCRUDAll)){
+            throw new BadRequestAlertException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         return produtoRepository.findAll();
     }
 
@@ -93,6 +118,10 @@ public class ProdutoResource {
     @GetMapping("/produtos/{id}")
     public ResponseEntity<Produto> getProduto(@PathVariable Long id) {
         log.debug("REST request to get Produto : {}", id);
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(UserAccountPermissionChecker.checkPermissao(contaTest, canCRUDAll)){
+            throw new BadRequestAlertException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         Optional<Produto> produto = produtoRepository.findById(id);
         return ResponseUtil.wrapOrNotFound(produto);
     }
@@ -106,6 +135,10 @@ public class ProdutoResource {
     @DeleteMapping("/produtos/{id}")
     public ResponseEntity<Void> deleteProduto(@PathVariable Long id) {
         log.debug("REST request to delete Produto : {}", id);
+        Conta contaTest = contaRepository.findByUserIsCurrentUser();
+        if(!contaTest.getNivelPermissao().equals(NivelPermissao.ADMIN)){
+            throw new BadRequestAlertException("Usuário não autorizado!", ENTITY_NAME, "not_authorized");
+        }
         produtoRepository.deleteById(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
